requirements.txt
appdirs==1.4.4
attrs==20.3.0
bcrypt==3.2.0
bleach==3.2.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.3.1
cycler==0.10.0
docutils==0.16
docx==0.2.4
docxtpl==0.11.2
future==0.18.2
google-api-core==1.24.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gspread==3.6.0
httplib2==0.18.1
idna==2.10
importlib-metadata==3.3.0
Jinja2==2.11.2
jsonschema==3.2.0
keyring==21.5.0
kiwisolver==1.3.1
lxml==4.6.2
markdown2==2.3.10
MarkupSafe==1.1.1
matplot==0.1.9
matplotlib==3.3.3
mutablerecords==0.4.1
ntlm-auth==1.5.0
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
openhtf-docx-report==0.0.3
packaging==20.8
paramiko==2.7.2
Pillow==8.0.1
ping3==2.6.6
pip-autoremove==0.9.1
pkginfo==1.6.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.3
pyloco==0.0.139
PyNaCl==1.4.0
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.4
pywin32-ctypes==0.2.0
PyYAML==5.3.1
readme-renderer==28.0
requests==2.25.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.6
SimpleWebSocketServer==0.1.1
six==1.15.0
sockjs-tornado==1.0.7
spintop-openhtf==0.6.5
SQLAlchemy==1.3.22
tenacity==8.0.1
tornado==4.5.3
tqdm==4.54.1
twine==3.2.0
typing==3.7.4.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
ushlex==0.99.1
webencodings==0.5.1
websocket-client==0.57.0
zipp==3.4.0
